\name{chromSizes}
\docType{data}
\alias{chromSizes}
\title{Information on the chromosome sizes}
\description{This dataset contains information on the lengths of the mouse chromosomes. This information was obtained using the fetchChromSizes tool from UCSC Genome Browser.}
\usage{chromSizes}
\format{A named vector containing lengths of the mouse chromosomes.}
\source{UCSC Genome Browser}
\references{\url{http://genome.ucsc.edu}}
\keyword{datasets}