\name{genePosition}
\docType{data}
\alias{genePosition}
\title{Information on genes positions}
\description{This dataset contains information on the position (chromosome, strand and TSS) of 25568 mouse genes as defined in Ensembl64 database.}
\usage{genePosition}
\format{A data.frame containing four columns (geneID, chromosome, strand, tss) and 25568 rows.}
\source{Ensembl64 via BioMart}
\references{\url{http://www.ensembl.org}, \url{http://www.biomart.org}}
\keyword{datasets}
