\name{peakPosition}
\docType{data}
\alias{peakPosition}
\title{Information on peaks positions}
\description{This dataset contains information on the position and intensity of 23774 peaks from the ChIP-seq experiment on the transcription factor CEBPa.}
\usage{peakPosition}
\format{A data.frame containing three colums (chromosome, center and intensity) and 23774 rows.}
% \source{}
\references{Hannah R. et al.
	\emph{A compendium of genome-wide hematopoietic transcription factor maps supports the identification of gene regulatory control mechanisms.} Exp Hematol. 2011 May;39(5):531-41}
\keyword{datasets}
